refactor(grpc): extract shared service helpers into conversions.go
Phase-2 gRPC cleanup. The proto<->model converters, pagination defaults,
and caller-identity helpers used by all six gRPC services were scattered
through secret_service.go and share_service.go, so those two files mixed
RPC handlers with package-wide plumbing. Moved the shared helpers
(requireUser, hasPermission, normalizePage, totalPages, secretNodeToProto,
sharingInfoToProto, metadataToMap, and the numeric/time/optional converters)
into a dedicated conversions.go. Each service file now holds only its own
RPC methods plus its service-specific error mapper.

Pure move — no behavior change (gRPC service tests unchanged and green).
secret_service.go 384→259, share_service.go 246→227, + conversions.go (172).

Note: the backlog suggested a finer CRUD/versions/listing split, but after
extracting the shared plumbing both files are appropriately sized
(~230-260 lines), so further sub-splitting would be over-engineering.
